requirements.txt
Django==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
Pillow==9.5.0
PyYAML==6.0